requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
psycopg2-binary
pytest
httpx
